+++ server/src/broadcast_msg.rs
@@ -3,5 +3,5 @@ use msgs::client_server_msg::Address;
 #[derive(Debug, Clone)]
 pub enum BroadcastMsg {
     Send (Address, Vec<u8>),
-    Kick (u32),
+    Kick (u16),
 }
